Treat applying a symbolic lambda as a constraint. #36
The main goal is to try and generate terms that will make
the application succeed or fail.

Updated the functional tests but there are some outputs that
ignore the spec that dictates an argument to be a function.
This will be fixed once the solver takes into account funs
in specs.