Use the correct arguments when evaluating a closure with a symbolic value #40
If the execution reaches a point where a closure associated with a
symbolic value, e.g. $1, is to be evaluated, then

1) it is a parameter, so we need to call it with the concrete parameters
to get the concrete value of the result. The symbolic value will be a
fresh symbolic variable, e.g. $3. In addition, we'll create a constraint
to represent the symbolic application. If the symbolic arguments are
$21, $22, $23, ... , $2n then we record the constraint
$3 = $1 ( $21, $22, $23, ...., $2n )

2) is it a closure that was created by our interpreter. In this
case we need to call it with the combined concrete and symbolic
arguments as it will decouple and use them properly internally.