Add support for 19.0
This involved changing one of the tests to use the new 'rand' module
insteed of the 'random' module which is marked deprecated in 19.0.
In fact, we still could support 17.x in CutEr if we were to be a bit
less strict in considering compiler warnings as errors. But I prefer
not to ignore warnings from later Erlang/OTP releases, so I drop
the 17.x releases from Travis and add 19.0 instead.