export smt with constraints to file
The set of constraints is described in SMT-LIB and printed to a file output.smt in current directory.
Output file is rewritten each time a constraint set is forwarded to the solver.
The SMT-LIB built-in type Bool is an alternative option for a Term. Atoms atomTrue and atomFalse are removed.
Currently only function g from module collection is supported.