Force function domains to contain at least one element (#149)
The pointwise definition of a function in the solver's logic allows a
function to have an empty domain set, which results in a total function
with an undefined value.

Forcing all functions to be defined for at least one point ensures that
there are no undefined values in the codomain.