Add a test case that currently challenges CutEr
The test case is currently not failing, although it should.
It has been put first on the ftests.json file so that it eases testing
when performing 'make ftest'.  This needs to be fixed: its solution
needs to be updated once it fails, and the json entry needs to be moved
to its proper place.

But it is an interesting test case nonetheless.
 - For starters, it shows that the printing of funs and function
   dependencies needs to be improved -- the third argument of the
   funs:ff_f_g/3 call needs to read A0 instead of A.
 - Also, the conjunctive call, once made properly should result in
   triggering the bug.