Do not attempt to fix variables to specific values
If the entry point has arity > 1, and the solve returns unknown or
timeout for a model, we would try to fix some of the variables to
specific values to simplify the model.

This functionality is not helpful in our case. If the solver reponds
with unknown or times out, then there is a regression in the solver or
in our tool.