Add a test that shows a regression
The added test is a small variation of a test that we already had in the
same test file.  I am sure it used to work OK in a previous version of
CutEr because I have been using it in demos.  Most likely, some recent
code refactoring broke it - perhaps we can use `git bisect` to find it.

What's weird is not so much that CutEr does not find the input that
crashes the unit under test, but that the inputs that we report are not
crashing inputs.