Appropriately handle funs as literals (#173)
The added test is a small variation of a test that we already had in the
same test file.  I am sure it used to work OK in a previous version of
CutEr because I have been using it in demos.

Turns out that, starting with Erlang/OTP 21.x, the compiler has started
treating known funs as literals and not create them with `erlang:make_fun/3`
as in previous versions.  Therefore, the `cuter_eval` module has been
appropriately extended to handle funs as literals.


Co-authored-by: Aggelos Giantsios <aggelgian@gmail.com>