Fix makefile when using a Python virtual environment
Under specific circumstances, the user-base directory reported
by the `python -m site --user-base` command does not
equal the directory where binaries are installed (e.g. in the
case of a Python virtual environment).

Update the configure script to actively look for the protoc-gen-mypy
binary under the Python sys.exec_prefix, and fallback to the old
path if needed. In any case, error out if the binary is not found.