Record PIDs and refs in log traces
Currently, we were dropping log entries with PIDs and refs,
which led to missing critical information for certain programs:
- when a process would send a message like {self(), 42}
- when a process would consume a message from the mailbox

The solver does not need to reason about these type, but just
"tolerate" them. We do not expect that they will be passed as
arguments to the unit under test, but they are being generated
as values during the execution of the program.